requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumbers==1.0.1
django-phonenumber-field==7
djangorestframework==3.12
requests==2.5
geopy==2.4
